+++ src/main/java/org/c4marathon/assignment/infra/config/QueryDslConfig.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.config;
+package org.c4marathon.assignment.infra.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
