+++ src/main/java/org/c4marathon/assignment/usecase/TransferUseCase.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.usecase;
 
 import java.util.concurrent.Callable;
 
-import org.c4marathon.assignment.config.property.AccountPolicyProperties;
+import org.c4marathon.assignment.infra.config.property.AccountPolicyProperties;
 import org.c4marathon.assignment.domain.service.mainaccount.MainAccountService;
 import org.c4marathon.assignment.domain.service.mainaccount.TransferService;
 import org.c4marathon.assignment.domain.validator.MainAccountValidator;
